fix: silence widening and precedence warnings in exec stack math
- cast EXEC_STACK_PAGES to uint64 before multiplying by PGSIZE
- apply cast at both user_stack_bottom and uvmalloc size sites

+++ kernel/core/exec.c
@@ -275,10 +275,11 @@ int execve_kernel(char *path, char argv[][PATH_MAX], int argc)
 	uint64 new_heap_bottom = sz + PGSIZE;
 	new_heap_top = sz + PGSIZE;
 	uint64 user_stack_top = PHYSTOP_LOW;
-	uint64 user_stack_bottom = PHYSTOP_LOW - EXEC_STACK_PAGES * PGSIZE;
+	uint64 user_stack_bottom =
+	    PHYSTOP_LOW - ((uint64) EXEC_STACK_PAGES * PGSIZE);
 
 	if (uvmalloc(user_pagetable, user_stack_bottom,
-		     EXEC_STACK_PAGES * PGSIZE, PTE_R | PTE_W) < 0) {
+		     (uint64) EXEC_STACK_PAGES * PGSIZE, PTE_R | PTE_W) < 0) {
 		goto bad;
 	}
